d05p02 (if you have 16+GB RAM and some time)

+++ 2023/05/main.h
@@ -1,8 +1,8 @@
 #ifndef MAIN_H_  
 #define MAIN_H_
 
-int taskone(char *lines[], int linecount);
-int tasktwo(char *lines[], int linecount);
+uint64_t taskone(char *lines[], int linecount);
+uint64_t tasktwo(char *lines[], int linecount);
 
 int convert(char *map[], int maplines, uint64_t *seeds, uint64_t arrSize);
